chore: convert all tasks to babel es6
