+++ .gulp/paths.js
@@ -6,7 +6,6 @@ const infos = require('../package.json')
   , today = new Date()
   , paths =  {
     'src': './sources',
-    'common': './sources/settings/commons.json',
     'utils': './utils',
     'schemes': './schemes',
     'widgets': './widgets',
