--- extras/PlainTasks/Material-Theme-Lighter.YAML-tmTheme
@@ -1,192 +1,0 @@
-# [PackageDev] target_format: plist, ext: hidden-tmTheme
-# > PlainTasks
----
-name: 'Material-Theme-Lighter Plain Tasks'
-uuid: 'adbb93d6-d3d2-4a08-82aa-8a5d163a83f9'
-author: 'Mattia Astorino'
-
-colorSpaceName: 'sRGB'
-
-semanticClass: 'material.theme.lighter.plain.tasks'
-settings:
-- settings:
-    background: '#fafafa'
-    caret: '#27272790'
-    foreground: '#80CBC4'
-    gutter: '#fafafa'
-    invisibles: ''
-    lineHighlight: '#90A4AE20'
-    selection: '#90A4AE20'
-    selectionBorder: '#90A4AE20'
-    guide: '#90A4AE20'
-    stackGuide: '#90A4AE20'
-    activeGuide: '#90A4AE20'
-
-# >> General
-
-- name: Comments
-  scope: comment
-  settings:
-    foreground: ''
-    fontStyle: 'italic'
-
-- name: Tags
-  scope: meta.tag.todo
-  settings:
-    foreground: '#945EB8'
-    fontStyle: 'bold'
-
-- name: Notes
-  scope: notes.todo
-  settings:
-    foreground: ''
-
-- name: Meta
-  scope: meta
-  settings:
-    fontStyle: 'italic'
-
-- name: Done Bullet
-  scope: punctuation.definition.bullet.completed
-  settings:
-    foreground: '#91B859'
-    fontStyle: 'bold'
-
-- name: Pending Bullet
-  scope: punctuation.definition.bullet.pending
-  settings:
-    foreground: '#80CBC4aa'
-
-- name: Cancelled Bullet
-  scope: punctuation.definition.bullet.cancelled
-  settings:
-    foreground: '#E53935'
-
-- name: Cancelled Task
-  scope: meta.item.todo.cancelled
-  settings:
-    foreground: ''
-    fontStyle: 'italic'
-
-- name: Meta Info
-  scope: meta.tag.todo.completed, meta.tag.todo.cancelled
-  settings:
-    foreground: 'aa'
-
-- name: Keyword
-  scope: keyword
-  settings:
-    background: '#90A4AE20'
-    foreground: '#80CBC4dd'
-    fontStyle: 'bold'
-
-- name: Separator
-  scope: meta.punctuation.archive.todo, meta.punctuation.separator.todo
-  settings:
-    foreground: ''
-    fontStyle: 'bold'
-
-- name: Font Style Punctuation
-  scope: punctuation.definition.italic, punctuation.definition.bold
-  settings:
-    foreground: 'aa'
-    fontStyle: 'normal'
-
-- name: Italic
-  scope: todo.italic
-  settings:
-    fontStyle: 'italic'
-
-- name: Bold
-  scope: todo.bold
-  settings:
-    fontStyle: 'bold'
-
-- name: URL
-  scope: todo.url
-  settings:
-    foreground: '#6182B8'
-    fontStyle: 'bold'
-
-- name: String
-  scope: string
-  settings:
-    background: '#FFB62Cbf'
-    foreground: '#000000bf'
-
-- name: Number
-  scope: constant.numeric
-  settings:
-    background: '#91B859bf'
-    foreground: '#000000bf'
-
-- name: Constant
-  scope: constant.language, constant.character, constant.other
-  settings:
-    background: '#6182B8bf'
-    foreground: '#000000bf'
-
-- name: Variable
-  scope: variable.language, variable.other
-  settings:
-    background: '#39ADB5bf'
-    foreground: '#000000bf'
-    fontStyle: ''
-
-- name: Entity
-  scope: constant.character.entity
-  settings:
-    background: '#91B859bf'
-    foreground: '#000000bf'
-
-- name: Storage
-  scope: storage
-  settings:
-    background: 'bf'
-    foreground: '#000000bf'
-
-- name: Class Name
-  scope: entity.name.type.class, support.class
-  settings:
-    foreground: '#6182B8bf'
-    fontStyle: ''
-
-- name: Tag Name
-  scope: entity.name.tag, entity.other.attribute-name.class.css, entity.other.attribute-name.id.css
-  settings:
-    background: '#FF5370bf'
-    foreground: '#000000bf'
-
-- name: Invalid
-  scope: invalid
-  settings:
-    foreground: '#ffffff'
-    background: '#E53935'
-
-- name: Today
-  scope: string.other.tag.todo.today
-  settings:
-    background: '#6182B8'
-    foreground: '#ffffff'
-    fontStyle: 'bold'
-
-- name: Low
-  scope: string.other.tag.todo.low
-  settings:
-    background: '#90A4AE20'
-    foreground: '#80CBC4cc'
-    fontStyle: 'bold'
-
-- name: High
-  scope: string.other.tag.todo.high
-  settings:
-    background: '#F76D47'
-    foreground: '#ffffff'
-    fontStyle: 'bold'
-
-- name: Critical
-  scope: string.other.tag.todo.critical
-  settings:
-    background: '#E53935'
-    foreground: '#ffffff'
-    fontStyle: 'bold'
