+++ package.json
@@ -11,6 +11,7 @@
   "devDependencies": {
     "babel-core": "^6.18.2",
     "babel-preset-es2015": "^6.18.0",
+    "babel-root-import": "^4.1.3",
     "colors": "^1.1.2",
     "conventional-github-releaser": "^1.1.2",
     "del": "^2.2.0",
@@ -31,6 +32,7 @@
     "gulp-replace": "^0.5.4",
     "gulp-template": "^4.0.0",
     "lodash": "^4.15.0",
+    "require-dir": "^0.3.1",
     "run-sequence": "^1.1.5",
     "sleep": "^5.0.0",
     "yargs": "^6.2.0"
